refactor: 회원 도메인 로직 수정 (#133)
