+++ src/main/kotlin/kr/wooco/woocobe/course/ui/web/controller/response/CourseCommentDetailResponse.kt
@@ -1,7 +1,7 @@
 package kr.wooco.woocobe.course.ui.web.controller.response
 
 import kr.wooco.woocobe.course.domain.model.CourseComment
-import kr.wooco.woocobe.user.domain.model.User
+import kr.wooco.woocobe.user.domain.entity.User
 import java.time.LocalDateTime
 
 data class CourseCommentDetailResponse(
@@ -40,8 +40,8 @@ data class CourseCommentWriterResponse(
         fun from(user: User): CourseCommentWriterResponse =
             CourseCommentWriterResponse(
                 id = user.id,
-                name = user.name,
-                profileUrl = user.profileUrl,
+                name = user.profile.name,
+                profileUrl = user.profile.profileUrl,
             )
     }
 }
