+++ src/main/kotlin/kr/wooco/woocobe/place/ui/web/controller/response/PlaceReviewDetailsResponse.kt
@@ -1,7 +1,7 @@
 package kr.wooco.woocobe.place.ui.web.controller.response
 
 import kr.wooco.woocobe.place.domain.model.PlaceReview
-import kr.wooco.woocobe.user.domain.model.User
+import kr.wooco.woocobe.user.domain.entity.User
 import java.time.LocalDateTime
 
 data class PlaceReviewDetailsResponse(
@@ -77,8 +77,8 @@ data class PlaceReviewWriterResponse(
         fun from(user: User): PlaceReviewWriterResponse =
             PlaceReviewWriterResponse(
                 id = user.id,
-                name = user.name,
-                profileUrl = user.profileUrl,
+                name = user.profile.name,
+                profileUrl = user.profile.profileUrl,
             )
     }
 }
