+++ src/main/kotlin/kr/wooco/woocobe/user/domain/gateway/UserStorageGateway.kt
@@ -1,6 +1,6 @@
 package kr.wooco.woocobe.user.domain.gateway
 
-import kr.wooco.woocobe.user.domain.model.User
+import kr.wooco.woocobe.user.domain.entity.User
 
 interface UserStorageGateway {
     fun save(user: User): User
