+++ src/main/kotlin/kr/wooco/woocobe/user/infrastructure/gateway/UserStorageGatewayImpl.kt
@@ -2,7 +2,7 @@ package kr.wooco.woocobe.user.infrastructure.gateway
 
 import kr.wooco.woocobe.user.domain.exception.NotExistsUserException
 import kr.wooco.woocobe.user.domain.gateway.UserStorageGateway
-import kr.wooco.woocobe.user.domain.model.User
+import kr.wooco.woocobe.user.domain.entity.User
 import kr.wooco.woocobe.user.infrastructure.storage.UserStorageMapper
 import kr.wooco.woocobe.user.infrastructure.storage.repository.UserJpaRepository
 import org.springframework.data.repository.findByIdOrNull
