+++ src/main/kotlin/kr/wooco/woocobe/user/infrastructure/storage/UserStorageMapper.kt
@@ -1,6 +1,6 @@
 package kr.wooco.woocobe.user.infrastructure.storage
 
-import kr.wooco.woocobe.user.domain.model.User
+import kr.wooco.woocobe.user.domain.entity.User
 import kr.wooco.woocobe.user.infrastructure.storage.entity.UserJpaEntity
 import org.springframework.stereotype.Component
 
@@ -9,16 +9,18 @@ class UserStorageMapper {
     fun toDomain(userJpaEntity: UserJpaEntity): User =
         User(
             id = userJpaEntity.id,
-            name = userJpaEntity.name,
-            profileUrl = userJpaEntity.profileUrl,
-            description = userJpaEntity.description,
+            profile = User.UserProfile(
+                name = userJpaEntity.name,
+                profileUrl = userJpaEntity.profileUrl,
+                description = userJpaEntity.description,
+            ),
         )
 
     fun toEntity(user: User): UserJpaEntity =
         UserJpaEntity(
             id = user.id,
-            name = user.name,
-            profileUrl = user.profileUrl,
-            description = user.description,
+            name = user.profile.name,
+            profileUrl = user.profile.profileUrl,
+            description = user.profile.description,
         )
 }
